old/requirements.txt
spidev
RPi.GPIO
spidev
RPi.GPIO
Pillow
smbus2
